update install script

+++ install.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
+# Set mish as a defualt shell
+
 INSTALL_DIR="/usr/bin"
 
 echo "Installing mish"
-cp ./mish "$INSTALL_DIR"
 
-if ! grep -Fxq "$INSTALL_DIR/mish" /etc/shells; then
-    echo "/usr/bin/mish" >> /etc/shells
+if [ -f "mish" ]; then
+    cp ./mish "$INSTALL_DIR"
+
+    if ! grep -Fxq "$INSTALL_DIR/mish" /etc/shells; then
+        echo "/usr/bin/mish" >> /etc/shells
+    fi
+else
+    echo "couldn't find executable"
 fi
